Inline the shared helpers
